requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu